Move ArtifactStoreState type to domain layer
Type definitions belong in domain per AGENTS.md; the classifier
function stays in business-logic. Renames the file from
artifact-store-state to classify-artifact-store to reflect what
remains in business-logic.
