+++ src/modules/artifacts/domain/artifact-store-state.ts
@@ -1,0 +1,5 @@
+export type ArtifactStoreState =
+	| { kind: "local"; uri: string }
+	| { kind: "remote-no-connector"; uri: string }
+	| { kind: "remote-ok" }
+	| { kind: "unknown" };
